gpufort/translator: Improve expr type resolution
... when checking if an expression is a tensor access
or function call before generating C code.

Now checking in the following order:

  1. Is explicitly declared array / implicitly declared array (via
          DIMENSION)?
  2. Is explicitly declared procedure?
  3. Is intrinsic?

If the expression cannot be associated with the any
of the above, it must be a variable from an ignored module,
or an external procedure, which will available only
at the linking stage. This check is still missing.